Culled some files
