+++ .gitignore
@@ -1,9 +1,14 @@
 node_modules
+package-lock.json
 .env
 .DS_Store
-demo/dump.rdb
-dump.rdbquell-client/yarn.lock
-quell-server/yarn.lock
-demo/yarn.lock
 .rdb
 .DS_STORE
+quell-server/package-lock.json
+quell-server/node_modules
+quell-client/package-lock.json
+quell-client/node_modules
+quell-extension/node_modules
+quell-extension/package-lock.json
+demo/package-lock.json
+demo/node_modules
