+++ quell-client/src/Quellify.js
@@ -1,15 +1,14 @@
 const { parse } = require('graphql/language/parser');
-const parseAST = require('./helpers/parseAST');
 const mapGenerator = require('./helpers/mapGenerator');
 const {
   lokiClientCache,
   normalizeForLokiCache,
 } = require('./helpers/normalizeForLokiCache');
-const { buildFromCache, generateCacheID } = require('./helpers/buildFromCache'); 
+const parseAST = require('./helpers/parseAST');
 const updateProtoWithFragment = require('./helpers/updateProtoWithFragments');
 
 // NOTE:
-// options feature is currently EXPERIMENTAL and the intention is to give Quell users the ability to customize cache update policies or to define custom IDs to use as keys when caching data
+// options feature is currently EXPERIMENTAL (and unused and has not been updated to work with LokiJS) and the intention is to give Quell users the ability to customize cache update policies or to define custom IDs to use as keys when caching data
 // keys beginning with __ are set aside for future development
 // defaultOptions provides default configurations so users only have to supply options they want control over
 const defaultOptions = {
@@ -25,37 +24,43 @@ const defaultOptions = {
   },
 };
 
+// used to generate Cache IDs to differentiate data in cache
+function generateCacheID(queryProto) {
+  // if ID field exists, set cache ID to 'fieldType--ID', otherwise just use fieldType
+  const cacheID = queryProto.__id
+    ? `${queryProto.__type}--${queryProto.__id}`
+    : queryProto.__type;
+
+  return cacheID;
+}
+
 /**
  * Quellify replaces the need for front-end developers who are using GraphQL to communicate with their servers
  * to write fetch requests. Quell provides caching functionality that a normal fetch request would not provide.
  * Quell syntax is similar to fetch requests and it includes the following:
  *    - accepts a user's endpoint and query string as inputs,
- *    - checks sessionStorage and constructs a response based on the requested information,
+ *    - checks LokiJS storage and constructs a response based on the requested information,
  *    - reformulates a query for any data not in the cache,
  *    - passes the reformulated query to the server to resolve,
  *    - joins the cached and server responses,
  *    - decomposes the server response and caches any additional data, and
  *    - returns the joined response to the user.
- *  @param {string} endPoint - The address to where requests are sent and processed. E.g. '/graphql'
+ *  @param {string} endPoint - The address to where requests are sent and processed. E.g. '/graphql'. This is also used to generate the maps (via mapGenerator) needed to process the query.
  *  @param {string} query - The graphQL query that is requested from the client
- *  @param {object} mutationMap - map of mutation that will be used to create mutation object and determine mutation type
- *  @param {object} map - JavaScript object with a key-value pair for every valid root query - type defined in the user's GraphQL schema
- *  @param {object} queryTypeMap - map of queryType that will be used when caching on client cache storage
  *  @param {object} userOptions - JavaScript object with customizable properties (note: this feature is still in development, please see defaultOptions for an example)
  */
 
-async function Quellify(endPoint, query, maps, userOptions = {}) {
+async function Quellify(endPoint, query, userOptions = {}) {
   // merge defaultOptions with userOptions
   // defaultOptions will supply any necessary options that the user hasn't specified
-
-  // mapGenerator is used to generate mutationMap, map and queryTypeMap, these were all required inputs prior to the latest revision.
-  const { map, queryTypeMap, mutationMap } = maps;
-
   const options = { ...defaultOptions, ...userOptions };
   let typeOfOperation = {
     isMutation: false,
     typeOfMutation: '',
   };
+
+  // mapGenerator is used to generate mutationMap, map and queryTypeMap
+  const { map, queryTypeMap, mutationMap } = mapGenerator(endPoint);
 
   // iterate over map to create all lowercase map for consistent caching
   for (const props in map) {
@@ -67,18 +72,6 @@ async function Quellify(endPoint, query, maps, userOptions = {}) {
 
   // Create AST based on the input query using the parse method available in the graphQL library (further reading: https://en.wikipedia.org/wiki/Abstract_syntax_tree)
   const AST = parse(query);
-
-  /**
-   * parseAST creates a proto object that contains a key for every root query in the user's request. Also, every root query key contains a key for each field requested on that root query, which is assigned the value of "true". The proto object also carries the following details for every root query
-   *    __args - arguments the user passed into the query (null if no arguments were given)
-   *    __alias - alias the user included in the query (null if no arguments were given)
-   *    __type - the type of root query as defined in the GraphQL schema, which could also be found in the map object passed into Quellify
-   *    __id - the ID assigned to the query, either by the user or by the database
-   * parseAST also creates an operationType that will evaluate to 'unQuellable' if the request is out-of-scope for caching (please see usage notes in the Readme.md for more details).
-   * parseAST also creates a frags object that contains any fields that were requested by a user who chose to use fragments in their request
-   *  @param {object} AST - Abstract Syntax Tree generated by parsing the input query (please see here for examples -> https://astexplorer.net/)
-   *  @param {object} options - JavaScript object defined in defaultOptions if not provided by the developer using Quell
-   */
 
   //create proto, operationType, and frags using parseAST
   const { proto, operationType, frags } = parseAST(AST, options);
@@ -100,18 +93,12 @@ async function Quellify(endPoint, query, maps, userOptions = {}) {
   } else if (operationType === 'mutation') {
     //if operationType is mutation
 
-    // ### DELETE ###
-    //if the mutationQuery is not coming from demo, mutation Query can be created using the code below
-    //let mutationQuery = createMutationStr(proto);
-    // ###        ###
 
     // create mutation object using mutationMap and proto created from parseAST;
     typeOfOperation.isMutation = true;
     let mutationObject;
 
-    //loops over the mutation map and checks if our proto ( an object key for every root query in the user's request)
-    //has a key that is found in mutation map
-    // we then add that key, value pair into the mutationObject.
+    //loops over the mutation map and checks if our proto ( an object key for every root query in the user's request) has a key that is found in mutation map. we then add that key, value pair into the mutationObject.
     for (let mutation in mutationMap) {
       if (proto.hasOwnProperty(mutation)) mutationObject = proto[mutation];
     }
@@ -177,35 +164,11 @@ async function Quellify(endPoint, query, maps, userOptions = {}) {
     }
   } else {
     // if the request is query
-
-    /**
-     * updateProtoWithFragment iterates over the fragments provied by a user and converts them into fields with values of true, and saves them to a new prototype object
-     *  @param {object} proto - JavaScript object generated by parseAST
-     *  @param {object} frags - JavaScript object with a key of the fragment name defined by the user, and properties for each field requested on that fragment
-     * check if the user's request included fragments before invoking updateProtoWithFragment
-     */
-
     const prototype =
       Object.keys(frags).length > 0
         ? updateProtoWithFragment(proto, frags)
         : proto;
-    // create an array of root queries on the prototype object so that we can differentiate between root queries and fields nested in a root query
     const prototypeKeys = Object.keys(prototype);
-
-
-    //### DELETE ### ? This appears to be unused, perhaps from session storage?
-    /**
-     *  @param {object} prototype - JavaScript object generated by parseAST (or updateProtoWithFragment if the user request included fragments)
-     *  @param {array} prototypeKeys - List of root queries requested by the user
-     */
-
-
-
-    // initialize a cacheHasData to false
-    let cacheHasData = false;
-
-
-    //
     let cacheID;
     let specificID;
     let actionQuery;
@@ -225,7 +188,6 @@ async function Quellify(endPoint, query, maps, userOptions = {}) {
 
 
     //if currField from Cache is an array , do below logic to get CacheIDArr
-
     let lokiJS = lokiClientCache.data;
     const cacheIDArr = [],
       cacheArr = [],
@@ -254,13 +216,8 @@ async function Quellify(endPoint, query, maps, userOptions = {}) {
       }
     });
 
-
-
-    if (cacheIDArr.length > 0) cacheHasData = true;
-    if (dataInLoki.length > 0) cacheHasData = true;
-
-    if (!cacheHasData) {
-
+    // checking if cache has data before continuing, if the data is not cached, query the database
+    if (!cacheIDArr.length > 0 || !dataInLoki.length > 0) {
       const fetchOptions = {
         method: 'POST',
         headers: options.headers,
@@ -280,7 +237,7 @@ async function Quellify(endPoint, query, maps, userOptions = {}) {
       // Return response as a promise
       return new Promise((resolve, reject) => resolve({ data: parsedData }));
     }
-
+    // SERVERSIDE: if cache has data, iterate through the data and create a cache object
     if (cacheIDArr.length > 0) {
       cacheIDArr.forEach((ID) => {
         let idx = 0;
@@ -297,7 +254,7 @@ async function Quellify(endPoint, query, maps, userOptions = {}) {
 
       return new Promise((resolve, reject) => resolve(cacheResponse));
     }
-
+    // CLIENTSIDE: if cache has data, iterate through the data and create a cache object
     if (dataInLoki.length > 0) {
       let cacheInfo = dataInLoki[0]['cacheID'];
       let info = { [`${actionQuery}`]: cacheInfo };
